pynvml conditional import added

+++ lib/galaxy/jobs/runners/local.py
@@ -28,11 +28,20 @@ import sys, signal
 import csv
 import shutil
 
-import pynvml as nvml
+gpu_flag = 0
+bash_command = "/bin/bash -c 'nvidia-smi'"
+sp = subprocess.Popen(bash_command, shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+out, err = sp.communicate()
+command_not_found = 'command not found'
+if command_not_found.encode() not in out:
+    gpu_flag = 1
+    import pynvml as nvml
 
 log = logging.getLogger(__name__)
-nvml.nvmlInit()
-gpu_count = nvml.nvmlDeviceGetCount()
+
+if gpu_flag == 1:
+    nvml.nvmlInit()
+    gpu_count = nvml.nvmlDeviceGetCount()
 
 __all__ = ('LocalJobRunner', )
 
@@ -73,7 +82,7 @@ class LocalJobRunner(BaseJobRunner):
             for req in reqmnts:
                 if req.type == "compute" and req.name == "gpu":
                     flag = 1
-            if gpu_count > 0 and flag == 1:
+            if gpu_flag == 1 and gpu_count > 0 and flag == 1:
                 # log.info("**************************CL  GPU ENABLED!!!!!**********************************************")
                 os.environ['GALAXY_GPU_ENABLED'] = "true"
             else:
@@ -208,7 +217,7 @@ class LocalJobRunner(BaseJobRunner):
             log.debug('({}) executing job script: {}'.format(job_id, command_line))
             
             #gpu_stats
-            if os.environ['GALAXY_GPU_ENABLED'] == "true":
+            if gpu_flag == 1 and os.environ['GALAXY_GPU_ENABLED'] == "true":
                 # log.info("**************************SUBMIT  GPU ENABLED!!!!!**********************************************")
                 directory = "gpu_util_%s" % tool_name
                 parent_dir = os.getcwd()
@@ -276,24 +285,24 @@ class LocalJobRunner(BaseJobRunner):
                 terminated = self.__poll_if_needed(proc, job_wrapper, job_id)
                 proc.wait()  # reap
                 if terminated:
-                    if os.environ['GALAXY_GPU_ENABLED'] == "true":
+                    if gpu_flag == 1 and os.environ['GALAXY_GPU_ENABLED'] == "true":
                         os.kill(sp,9)
                         self.post_process(log_file_path, util_path, util_mem_path, mem_free_path, mem_used_path, pcie_link_gen_cur_path, stats_path)
                     return
                 elif check_pg(proc.pid):
-                    if os.environ['GALAXY_GPU_ENABLED'] == "true":
+                    if gpu_flag == 1 and os.environ['GALAXY_GPU_ENABLED'] == "true":
                         os.kill(sp,9)
                         self.post_process(log_file_path, util_path, util_mem_path, mem_free_path, mem_used_path, pcie_link_gen_cur_path, stats_path)
                     kill_pg(proc.pid)
             finally:
                 with self._proc_lock:
-                    if os.environ['GALAXY_GPU_ENABLED'] == "true":
+                    if gpu_flag == 1 and os.environ['GALAXY_GPU_ENABLED'] == "true":
                         os.kill(sp,9)
                         self.post_process(log_file_path, util_path, util_mem_path, mem_free_path, mem_used_path, pcie_link_gen_cur_path, stats_path)
                     self._procs.remove(proc)
 
             if proc.terminated_by_shutdown:
-                if os.environ['GALAXY_GPU_ENABLED'] == "true":
+                if gpu_flag == 1 and os.environ['GALAXY_GPU_ENABLED'] == "true":
                     os.kill(sp,9)
                     self.post_process(log_file_path, util_path, util_mem_path, mem_free_path, mem_used_path, pcie_link_gen_cur_path, stats_path)
                 self._fail_job_local(job_wrapper, "job terminated by Galaxy shutdown")
@@ -306,11 +315,11 @@ class LocalJobRunner(BaseJobRunner):
             stdout_file.close()
             stderr_file.close()
             log.debug('execution finished: %s' % command_line)
-            if os.environ['GALAXY_GPU_ENABLED'] == "true":
+            if gpu_flag == 1 and os.environ['GALAXY_GPU_ENABLED'] == "true":
                 os.kill(sp,9)
                 self.post_process(log_file_path, util_path, util_mem_path, mem_free_path, mem_used_path, pcie_link_gen_cur_path, stats_path)
         except Exception:
-            if os.environ['GALAXY_GPU_ENABLED'] == "true":
+            if gpu_flag == 1 and os.environ['GALAXY_GPU_ENABLED'] == "true":
                 os.kill(sp,9)
                 self.post_process(log_file_path, util_path, util_mem_path, mem_free_path, mem_used_path, pcie_link_gen_cur_path, stats_path)
             log.exception("failure running job %d", job_wrapper.job_id)
